engage-ui fixed build issue for library publish (#16)
* #29820 worked on fixing build issues

* removed ref

* removed section and pane
