+++ src/components/ButtonGroup/ButtonGroup.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { themr, ThemedComponentClass } from 'react-css-themr';
 import {classNames} from '@shopify/react-utilities/styles';
 import {elementChildren} from '@shopify/react-utilities/components';
-import {Props as ButtonProps} from '../Button';
+// import {Props as ButtonProps} from '../Button';
 import Item from './Item';
 
 import { BUTTON_GROUP } from '../ThemeIdentifiers';
@@ -10,7 +10,7 @@ import * as baseTheme from './ButtonGroup.scss';
 
 export interface Props {
   segmented?: boolean,
-  children?: React.ReactElement<ButtonProps>[],
+  children?: React.ReactNode,
   theme?: any,
 }
