+++ src/components/Choice/Choice.tsx
@@ -14,7 +14,7 @@ export interface Props {
   label: string,
   error?: Error,
   labelHidden?: boolean,
-  children?: string,
+  children?: React.ReactNode,
   helpText?: React.ReactNode,
   theme?: any,
 }
