--- src/components/Picker/IPickerInfo.ts
@@ -1,7 +1,0 @@
-export interface IPickerInfo {
-    Id?: number,
-    Name: string,
-    Description: string,
-    ImageUrl?: string,
-    Url?: string,
-}
