--- src/components/Picker/IPickerSource.ts
@@ -1,3 +1,0 @@
-export interface IPickerSource<T> {
-    performFilter(filterString: string): Promise<T[]>,
-}
