+++ src/components/Picker/Picker.tsx
@@ -3,8 +3,6 @@ import { themr, ThemedComponentClass } from 'react-css-themr';
 import { PICKER } from '../ThemeIdentifiers';
 import TextField from '../TextField';
 import { DisplayMoreInfo } from './PickerEnum';
-import { IPickerInfo } from './IPickerInfo';
-import { IPickerSource } from './IPickerSource';
 
 import * as baseTheme from './Picker.scss';
 
@@ -13,6 +11,16 @@ export interface State {
     searchItems: IPickerInfo[],
     selectedItems: IPickerInfo[],
     moreInfo: boolean,
+}
+export interface IPickerInfo {
+    Id?: number,
+    Name: string,
+    Description: string,
+    ImageUrl?: string,
+    Url?: string,
+}
+export interface IPickerSource<T> {
+    performFilter(filterString: string): Promise<T[]>,
 }
 export type Type = 'hide' | 'mark';
 export interface Props {
