+++ src/components/ValidatedForm/ValidatedForm.tsx
@@ -1,5 +1,3 @@
-/// <reference path="../../../@types/rc-form.d.ts" />
-
 import * as React from 'react';
 import { createForm } from 'rc-form';
