+++ tsconfig.json
@@ -33,7 +33,8 @@
   },
   "include": [
     "./config/typescript/**/*",
-    "./src/**/*"
+    "./src/**/*",
+    "./typings/*"
   ],
   "exclude": [
     "./@types"
